Huge refactoring. Moved all reflection logic out
Now has a dedicated class to do the reflection
support for interfaces
no longer required to implement IController
added (early) support for ServiceContracts
support multiple methods with the same name, as long as they do not have
the same http verb signature. This probably needs a bit more work
because a method with Get & Post is not currently the same as a method
with the same name that has Post & Put, but Mvc will not like this.
UI change on the html view.
give areas with "no name" a "default" name